feat: enhance ExternalClientOptions and add extensions
Update the `ExternalClientOptions` class to include XML documentation
comments, modify the `Timeout` property with a new default value of 10
seconds, and add a `DefaultRequestHeaders` property. Introduce the
`ApplyTo` method for configuring an `HttpClient` instance.

Add the `ExternalClientServiceCollectionExtensions` class to provide
extension methods for adding and configuring external HTTP clients in
the service collection. The `AddExternalClient` method facilitates
registration of external HTTP clients with validation of options.

Related issues: #1